Rename `src/lib/import-export/types.ts`
The file exports a function so it shouldn't be named `types`.
